requirements.txt
fastapi
uvicorn
sentence-transformers
groq
faiss-cpu
numpy
tenacity
deepeval
pytest